+++ routes/department.route.js
@@ -24,16 +24,23 @@ router.route('/department/:department_id')
       });
     })
     .delete(function(req, res, next) {
-      Department.findByIdAndRemove(req.params.department_id).then(function(removedDocument) {
-        if (removedDocument) {
-          res.sendStatus(204);
-          winston.info(`Removed department with id ${req.params.department_id}`);
+      Program.find({department: req.params.department_id}).then(function(dependents) {
+        if (dependents.length === 0) {
+          Department.findByIdAndRemove(req.params.department_id).then(function(removedDocument) {
+            if (removedDocument) {
+              res.sendStatus(204);
+              winston.info(`Removed department with id ${req.params.department_id}`);
+            } else {
+              res.sendStatus(404);
+              winston.info(`Tried to remove nonexistent department with id ${req.params.department_id}`);
+            }
+          }, function(err) {
+            next(err);
+          });
         } else {
-          res.sendStatus(404);
-          winston.info(`Tried to remove nonexistent department with id ${req.params.department_id}`);
+          res.sendStatus(400);
+          winston.info(`Tried to remove department with id ${req.params.department_id} but it had dependents`);
         }
-      }, function(err) {
-        next(err);
       });
     });
